Extract BlockCipherTransform base class and split test hierarchy
All three block cipher transforms (ThreefishTransform, SkipjackTransform,
BlowfishTransform) contained identical implementations of TransformBlock,
TransformFinalBlock, Dispose, Combine, and all ICryptoTransform properties.

Extract the shared logic into an abstract BlockCipherTransform base class.
Each concrete transform is now a thin sealed class that delegates to base via
its constructor. BlowfishTransform retains its internal constructor visibility.

Split the test hierarchy to keep CryptoTransformTests small (only disposal
contract tests) since HashAlgorithm tests also inherit from it:

  CryptoTransformTests<TCryptoTransform>        — ICryptoTransform disposal
    └─ BlockCipherTransformTests<TCryptoTransform> — null-arg _fix tests,
       │                                            property invariants
       ├─ ThreefishTransformTests
       ├─ SkipjackTransformTests
       └─ BlowfishTransformTests
