+++ Bodu.Security.Cryptography/test/Security.Cryptography/BlowfishTransformTests.cs
@@ -7,12 +7,12 @@
 namespace Bodu.Security.Cryptography
 {
     /// <summary>
-    /// Concrete test class that exercises the <see cref="CryptoTransformTests{TCryptoTransform}" /> base tests
+    /// Concrete test class that exercises the <see cref="BlockCipherTransformTests{TCryptoTransform}" /> base tests
     /// against the <see cref="BlowfishTransform" /> implementation.
     /// </summary>
     [TestClass]
     public sealed class BlowfishTransformTests
-        : CryptoTransformTests<BlowfishTransform>
+        : BlockCipherTransformTests<BlowfishTransform>
     {
         /// <inheritdoc />
         protected override BlowfishTransform CreateTransform()
